Manages invoice financing for supply chain finance
