+++ tests/letter-of-credit.test.ts
@@ -1,174 +1,21 @@
-import { describe, it, expect, beforeEach } from "vitest"
-import { simnet, Tx, types } from "your-library-name" // Replace "your-library-name" with the actual library name
 
-describe("Letter of Credit Contract", () => {
-  let accounts
-  let bankAddress
-  let beneficiaryAddress
-  let applicantAddress
-  let managerAddress
-  
-  beforeEach(() => {
-    accounts = {
-      deployer: "ST1PQHQKV0RJXZFY1DGX8MNSNYVE3VGZJSRTPGZGM",
-      wallet_1: "ST2CY5V39NHDPWSXMW9QDT3HC3GD6Q6XX4CFRK9AG",
-      wallet_2: "ST2JHG361ZXG51QTKY2NQCVBPPRRE2KZB1HR05NNC",
-      wallet_3: "ST2NEB84ASENDXKYGJPQW86YXQCEFEX2ZQPG87ND",
-    }
-    bankAddress = accounts.deployer
-    beneficiaryAddress = accounts.wallet_1
-    applicantAddress = accounts.wallet_2
-    managerAddress = accounts.wallet_3
-  })
-  
-  it("should verify LC manager", () => {
-    // Test manager verification
-    const block = simnet.mineBlock([
-      Tx.contractCall(
-          "trade-finance-manager",
-          "verify-manager",
-          [types.principal(managerAddress), types.ascii("Manager One"), types.ascii("MGR001"), types.uint(2000)],
-          bankAddress,
-      ),
-    ])
-    
-    expect(block.receipts[0].result).toBeOk(types.bool(true))
-  })
-  
-  it("should create a letter of credit", () => {
-    // Test LC creation after manager verification
-    simnet.mineBlock([
-      Tx.contractCall(
-          "trade-finance-manager",
-          "verify-manager",
-          [types.principal(managerAddress), types.ascii("Manager One"), types.ascii("MGR001"), types.uint(2000)],
-          bankAddress,
-      ),
-    ])
-    
-    const block = simnet.mineBlock([
-      Tx.contractCall(
-          "letter-of-credit",
-          "create-letter-of-credit",
-          [
-            types.principal(beneficiaryAddress),
-            types.principal(applicantAddress),
-            types.uint(100000),
-            types.ascii("USD"),
-            types.uint(1500),
-            types.principal(managerAddress),
-          ],
-          bankAddress,
-      ),
-    ])
-    
-    expect(block.receipts[0].result).toBeOk(types.uint(1))
-  })
-  
-  it("should issue a letter of credit", () => {
-    // Test LC issuance
-    simnet.mineBlock([
-      Tx.contractCall(
-          "trade-finance-manager",
-          "verify-manager",
-          [types.principal(managerAddress), types.ascii("Manager One"), types.ascii("MGR001"), types.uint(2000)],
-          bankAddress,
-      ),
-    ])
-    
-    simnet.mineBlock([
-      Tx.contractCall(
-          "letter-of-credit",
-          "create-letter-of-credit",
-          [
-            types.principal(beneficiaryAddress),
-            types.principal(applicantAddress),
-            types.uint(100000),
-            types.ascii("USD"),
-            types.uint(1500),
-            types.principal(managerAddress),
-          ],
-          bankAddress,
-      ),
-    ])
-    
-    const block = simnet.mineBlock([
-      Tx.contractCall("letter-of-credit", "issue-letter-of-credit", [types.uint(1)], bankAddress),
-    ])
-    
-    expect(block.receipts[0].result).toBeOk(types.bool(true))
-  })
-  
-  it("should present documents", () => {
-    // Test document presentation
-    simnet.mineBlock([
-      Tx.contractCall(
-          "trade-finance-manager",
-          "verify-manager",
-          [types.principal(managerAddress), types.ascii("Manager One"), types.ascii("MGR001"), types.uint(2000)],
-          bankAddress,
-      ),
-    ])
-    
-    simnet.mineBlock([
-      Tx.contractCall(
-          "letter-of-credit",
-          "create-letter-of-credit",
-          [
-            types.principal(beneficiaryAddress),
-            types.principal(applicantAddress),
-            types.uint(100000),
-            types.ascii("USD"),
-            types.uint(1500),
-            types.principal(managerAddress),
-          ],
-          bankAddress,
-      ),
-    ])
-    
-    simnet.mineBlock([Tx.contractCall("letter-of-credit", "issue-letter-of-credit", [types.uint(1)], bankAddress)])
-    
-    const block = simnet.mineBlock([
-      Tx.contractCall(
-          "letter-of-credit",
-          "present-documents",
-          [types.uint(1), types.ascii("Invoice-001,BOL-002")],
-          beneficiaryAddress,
-      ),
-    ])
-    
-    expect(block.receipts[0].result).toBeOk(types.bool(true))
-  })
-  
-  it("should get letter of credit details", () => {
-    // Test getting LC details
-    simnet.mineBlock([
-      Tx.contractCall(
-          "trade-finance-manager",
-          "verify-manager",
-          [types.principal(managerAddress), types.ascii("Manager One"), types.ascii("MGR001"), types.uint(2000)],
-          bankAddress,
-      ),
-    ])
-    
-    simnet.mineBlock([
-      Tx.contractCall(
-          "letter-of-credit",
-          "create-letter-of-credit",
-          [
-            types.principal(beneficiaryAddress),
-            types.principal(applicantAddress),
-            types.uint(100000),
-            types.ascii("USD"),
-            types.uint(1500),
-            types.principal(managerAddress),
-          ],
-          bankAddress,
-      ),
-    ])
-    
-    const result = simnet.callReadOnlyFn("letter-of-credit", "get-letter-of-credit", [types.uint(1)], bankAddress)
-    
-    expect(result.result).toBeSome()
-  })
-})
+import { describe, expect, it } from "vitest";
+
+const accounts = simnet.getAccounts();
+const address1 = accounts.get("wallet_1")!;
+
+/*
+  The test below is an example. To learn more, read the testing documentation here:
+  https://docs.hiro.so/stacks/clarinet-js-sdk
+*/
+
+describe("example tests", () => {
+  it("ensures simnet is well initalised", () => {
+    expect(simnet.blockHeight).toBeDefined();
+  });
+
+  // it("shows an example", () => {
+  //   const { result } = simnet.callReadOnlyFn("counter", "get-counter", [], address1);
+  //   expect(result).toBeUint(0);
+  // });
+});
